+++ src/main/java/com/alper/worldcup/service/LeaderboardTimelineSeries.java
@@ -2,5 +2,5 @@ package com.alper.worldcup.service;
 
 import java.util.List;
 
-public record LeaderboardTimelineSeries(String username, String label, List<Long> data, String color) {
+public record LeaderboardTimelineSeries(String username, String label, List<Double> data, String color) {
 }
